requirements.txt
Django==1.6.1
djangorestframework==2.3.12
pytz==2014.10
Unipath==1.0
